Agregar funcion eliminar producto
