+++ Sistema de Gestion Veterinaria.dev
@@ -29,7 +29,7 @@ IncludeVersionInfo=0
 SupportXPThemes=0
 CompilerSet=0
 CompilerSettings=0;0;0;0;0;0;0;0;0;0;0;0;0;0;0;0;0;0;0;0;0;0;20;0;0;0
-UnitCount=11
+UnitCount=13
 
 [VersionInfo]
 Major=1
@@ -152,6 +152,26 @@ BuildCmd=
 
 [Unit11]
 FileName=Mascota.cpp
+CompileCpp=1
+Folder=
+Compile=1
+Link=1
+Priority=1000
+OverrideBuildCmd=0
+BuildCmd=
+
+[Unit12]
+FileName=inventario.h
+CompileCpp=1
+Folder=
+Compile=1
+Link=1
+Priority=1000
+OverrideBuildCmd=0
+BuildCmd=
+
+[Unit13]
+FileName=inventario.cpp
 CompileCpp=1
 Folder=
 Compile=1
