+++ inventario.h
@@ -1,0 +1,7 @@
+#ifndef INVENTARIO_H
+#define INVENTARIO_H
+
+void RegistrarProducto();
+void mostrarInventario();
+void eliminarInventario();
+#endif
