+++ menu.h
@@ -5,14 +5,9 @@ void Menu();
 
 //Oscar
 void menuPropietario();
-void eliminarPropietario();
-void actualizarPropietario();
 
 //Yair
 void menuInventario();
-void RegistrarProducto();
-void mostrarInventario();
-
 
 //wesley
 void menuMascotas();
